break out into better structure
